Refactor: Move checkInputsForDepositTxRequest to DepositTxValidation
Move test mock utils to TradeValidationTestUtils
